conf: Add a smmuv3Accel IOMMU model
Add support for specifying "smmuv3Accel" as the IOMMU model. In the following
commits, when the "smmuv3Accel" IOMMU model is parsed from the VM definition,
PXB controllers each associated with a SMMU node will be auto-added to the VM
definition, and VFIO devices will be routed to PXB controllers based on their
association with host SMMU nodes. VFIO devices with unique host SMMU node
associations will be attached to unique PXB controllers.

Signed-off-by: Nathan Chen <nathanc@nvidia.com>