conf: Add a nestedSmmuv3 IOMMU model
Add support for specifying "nestedSmmuv3" as the IOMMU model. In the following
commits, when the "nestedSmmuv3" IOMMU model is parsed from the VM definition,
PXB controllers and "nestedSmmuv3" devices will be auto-added to the VM
definition and VFIO devices will be routed to PXB controllers based on their
association with host SMMU nodes. The auto-added "nestedSmmuv3" devices will
each be attached to a PXB controller so that VFIO devices with unique host
SMMU node associations are attached to unique PXB controllers.

Signed-off-by: Nathan Chen <nathanc@nvidia.com>