qemu: Add test case for specifying multiple nested SMMUs
Implement a sub-test in qemuxmlconftest that mocks the scanning of sysfs to
determine how many "nestedSmmuv3" devices to assign to the VM when
"nestedSmmuv3" IOMMU model is parsed from the VM definition. Add a VM
definition example with an associated qemu command line that exercises the
detection of "nestedSmmuv3" IOMMU model and auto-creation of PXB controllers and
"nestedSmmuv3" devices.

Signed-off-by: Nathan Chen <nathanc@nvidia.com>