cover-letter: Route hostdevs to multiple nested SMMUs
Hi,

This is a draft solution for supporting multiple vSMMU instances
in a qemu VM.

Based on discussions/suggestions received for a previous RFC by
Nicolin here[0], the association of vSMMUs to VFIO devices
in VM PCIe topology should be moved out of qemu into libvirt.
In addition, the nested SMMU nodes should be passed to qemu as
pluggable devices.

To address these changes, this patch series introduces a new
"nestedSmmuv3" IOMMU model and "nestedSmmuv3" device type.
Upon specifying the nestedSmmuv3 IOMMU model, nestedSmmuv3
devices will be auto-added to the VM definition based on the
available SMMU nodes in the host's sysfs. The nestedSmmuv3
devices will each be attached to a separate PXB controller, and
VFIO devices will be routed to PXBs based on their association
with host SMMU nodes. This will maintain a VM PCIe topology
that allows for multiple nested SMMUs per Nicolin's original
qemu patch series in [0] and Shameer's work in [1] to remove
VM topology changes from qemu and allow the nested SMMUs to be
specified as pluggable devices.

TODO:
- No DMA mapping can found by UEFI when specifying multiple
passthrough devices in the VM definition, and VM boot is
subsequently blocked. We need to investigate this for the
next revision, but we don't encounter this issue when passing
through a single device.

- Shameer's qemu branch specifies nestedSmmuv3 bus number
with "pci-bus" instead of "bus", so the libvirt compilation
test args and qemu args in qemuBuildPCINestedSmmuv3DevProps()
need to be modified to match this revision of qemu. It will
be reverted to using "bus" in the next qemu revision.

This series is on Github:
https://github.com/NathanChenNVIDIA/libvirt/tree/nested-smmuv3-iommufd-12-02-24

Thanks,
Nathan

[0] https://lore.kernel.org/qemu-devel/cover.1719361174.git.nicolinc@nvidia.com/
[1] https://lore.kernel.org/qemu-devel/20241108125242.60136-1-shameerali.kolothum.thodi@huawei.com/