conf: Add an iommufd device type
Add support for an "iommufd" device type that will be translated to a qemu
"-object iommufd" command line argument. This device type has an "id"
attribute to specify the iommufd object id that can be associated with a
passthrough device, and an "fd" attribute to specify the fd for externally
opening the /dev/iommu and VFIO cdev by a management layer.

Signed-off-by: Nathan Chen <nathanc@nvidia.com>