cover letter: qemu: Implement support for iommufd and multiple vSMMUs
Hi,

This is a follow up to the first RFC patchset [0] for supporting multiple
vSMMU instances in a qemu VM. This patchset also introduces support for
using iommufd to propagate DMA mappings to kernel for assigned devices.

This patchset implements a new 'iommufd' device type to represent the
iommufd object in qemu command line, and implements new 'iommufdId' and
'iommufdFd' attributes for hostdev devices to be associated with the
iommufd object.

For instance, specifying the iommufd object and associated hostdev in a
VM definition:

  <devices>
    <hostdev mode='subsystem' type='pci' managed='yes'>
      <driver name='vfio'/>
      <source>
        <address domain='0x0000' bus='0x06' slot='0x12' function='0x2'/>
      </source>
      <iommufdId>iommufd0</iommufdId>
      <address type='pci' domain='0x0000' bus='0x00' slot='0x03' function='0x0'/>
    </hostdev>
    <iommufd>
      <id>iommufd0</id>
    </iommufd>
  </devices>

This would get translated to a qemu command line with the arguments below:

-object '{"qom-type":"iommufd","id":"iommufd0"}' \
-device '{"driver":"vfio-pci","host":"0000:06:12.2","id":"hostdev1","iommufd":"iommufd0","bus":"pci.0","addr":"0x3"}' \

If users would like to leverage qemu's iommufd feature to open the VFIO
cdev and /dev/iommu via an external management layer, the fd can be
specified like so in the VM definition:

  <devices>
    <hostdev mode='subsystem' type='pci' managed='yes'>
      <driver name='vfio'/>
      <source>
        <address domain='0x0000' bus='0x06' slot='0x12' function='0x2'/>
      </source>
      <iommufdId>iommufd0</iommufdId>
      <iommufdFd>23</iommufdFd>
      <address type='pci' domain='0x0000' bus='0x00' slot='0x03' function='0x0'/>
    </hostdev>
    <iommufd>
      <id>iommufd0</id>
      <fd>22</fd>
    </iommufd>
  </devices>

This would get translated ot a qemu command line with the arguments below:

-object '{"qom-type":"iommufd","id":"iommufd0","fd":"22"}' \
-device '{"driver":"vfio-pci","host":"0000:06:12.2","id":"hostdev1","iommufd":"iommufd0","fd":"23","bus":"pci.0","addr":"0x3"}' \

TODO:
- I updated the namespace and cgroup configuration to allow access to iommufd
paths at /dev/vfio/devices/vfio* and /dev/iommu. However, qemu needs to be
launched with user and group set to 'root' in order for these paths to be
accessible. A passthrough device represented by /dev/vfio/18 normally has
'root' user and group permissions but in the mount namespace it's changed to
'libvirt-qemu' and 'kvm'. I wasn't able to discern where this is happening by
looking at src/qemu/qemu_namespace.c and src/qemu/qemu_cgroup.c. Would you have
any pointers on how to change the iommufd paths' user and group permissions in
the libvirt mount namespace?

This series is on Github:
https://github.com/NathanChenNVIDIA/libvirt/tree/nested-smmuv3-iommufd-12-09-24

Thanks,
Nathan

[0] https://lists.libvirt.org/archives/list/devel@lists.libvirt.org/thread/7GDT6RX5LPAJMPP4ZSC4ACME6GVMG236/

Signed-off-by: Nathan Chen <nathanc@nvidia.com>