qemu: Add test case for specifying iommufd
Implement a sub-test in qemuxmlconftest that takes an iommufd attribute
from the iommu device stanza, and iommufdId and iommufdFd hostdev
attributes in a VM definition. Translate the VM definition to a qemu
command line that associates the hostdev withan iommufd object, as well
as a fd for externally opening /dev/iommu and VFIO cdev from an external
management layer.

Signed-off-by: Nathan Chen <nathanc@nvidia.com>