conf: Validate CMDQv setting
Use of the CMDQv feature requires the VM to be backed by contiguous memory
at the host level. Until we implement vEGM support in libvirt, insert a
check that fails enabling CMDQv if hugepages VM backing is not configured.

Insert a check that fails enabling CMDQv if SMMUv3 or nested SMMUv3 IOMMU
models are not specified, since CMDQv is meant to extend SMMUv3 instance's
command queue feature for virtualization usage.

Signed-off-by: Nathan Chen <nathanc@nvidia.com>