conf: Add an iommufd member struct to virDomainIOMMUDef
Add support for an "iommufd" virDomainIOMMUDef member that will be
translated to a qemu "-object iommufd" command line argument. This
iommufd struct has an "id" member to specify the iommufd object id
that can be associated with a passthrough device, and an "fd"
member to specify the fd for externally opening the /dev/iommu and
VFIO cdev by a management layer.

Signed-off-by: Nathan Chen <nathanc@nvidia.com>